requirements.txt
langchain
langchain-groq
langchain-community
langchain-huggingface
langgraph
faiss-cpu
sentence-transformers
python-dotenv
pydantic
